+++ src/main/java/org/zaproxy/zap/extension/hud/tutorial/pages/IntroPage.java
@@ -19,7 +19,6 @@
  */
 package org.zaproxy.zap.extension.hud.tutorial.pages;
 
-import org.zaproxy.zap.extension.hud.ExtensionHUD.Telemetry;
 import org.zaproxy.zap.extension.hud.tutorial.TutorialPage;
 import org.zaproxy.zap.extension.hud.tutorial.TutorialProxyServer;
 
@@ -34,11 +33,5 @@ public class IntroPage extends TutorialPage {
     @Override
     public String getName() {
         return NAME;
-    }
-
-    @Override
-    public String getHtml() {
-        this.getTutorialProxyServer().telemetryPoint(Telemetry.TUTORIAL_START);
-        return super.getHtml();
     }
 }