+++ CHANGELOG.md
@@ -5,10 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+- Use stats instead of bitly for telemetry
 
 ## [0.14.0] - 2022-09-22
-
 ### Added
 - Added accessibility fixes for rapid navigation between panels. 
 - Added semantic roles and better labelling for screen reader users.